requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-groq>=0.1.0
duckduckgo-search>=6.2.0
streamlit>=1.36.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

